pyproject.toml
[project]
name = "style-rag"
version = "3.12.0"
description = "独立RAG系统 - 用于风格学习和创作优化 / Independent RAG System for Style Learning and Creative Enhancement"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "AI Gen Novel", email = "example@example.com"}
]
keywords = ["rag", "embedding", "vector-database", "nlp", "chinese"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]
reranker = [
    "FlagEmbedding>=1.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
all = [
    "style-rag[server,reranker,dev]",
]

[project.scripts]
style-rag-index = "style_rag.cli.index_articles:main"
style-rag-query = "style_rag.cli.query:main"
style-rag-server = "style_rag.api.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["style_rag"]
